% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_palette.R
\name{categorical_colours}
\alias{categorical_colours}
\title{List of colour codes in the 'categorical' colour palette}
\usage{
categorical_colours(code = c("hex", "rgb"))
}
\arguments{
\item{code}{type of colour code required. Choices are:
\code{hex}, \code{rgb}}
}
\value{
vector of the required type of colour code  from the categorical palette
}
\description{
List of colour codes in the 'categorical' colour palette
}
