% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duo_palette.R
\name{duo_colours}
\alias{duo_colours}
\title{List of colour codes in the 'duo' colour palette}
\usage{
duo_colours(code = c("hex", "rgb"))
}
\arguments{
\item{code}{type of colour code required. Choices are:
\code{hex}, \code{rgb}}
}
\value{
vector of the required type of colour code  from the duo palette
}
\description{
List of colour codes in the 'duo' colour palette
}
