% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_picker.R
\name{palette_picker}
\alias{palette_picker}
\title{Palette selection and return}
\usage{
palette_picker(type, code = "hex")
}
\arguments{
\item{type}{type of palette required (categorical, duo, sequential, focus)}

\item{code}{type of colour code required (hex, rgb, cymk, names), default is hex}
}
\value{
vector of the required type of colour code  from the required palette
}
\description{
Used in the af_colours function. Returns a vector of the required colour codes from the palette selected in the input "type"
}
