% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_palette.R
\name{sequential_colours}
\alias{sequential_colours}
\title{List of colour codes in the 'sequential' colour palette}
\usage{
sequential_colours(code = c("hex", "rgb", "cymk", "names"))
}
\arguments{
\item{code}{type of colour code required. Choices are:
\code{hex}, \code{rgb}, \code{cymk}, \names{names}}
}
\value{
vector of the required type of colour code  from the sequential palette
}
\description{
List of colour codes in the 'sequential' colour palette
}
