% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_colours.R
\name{af_colours}
\alias{af_colours}
\title{Analysis Function colour palette function}
\usage{
af_colours(
  type = c("categorical", "duo", "sequential", "focus"),
  n,
  code = "hex"
)
}
\arguments{
\item{type}{Name of required palette. Choices are:
\code{categorical}, \code{duo},  \code{sequential},
\code{focus}}

\item{n}{Number of colours to return, required for categorical palette type.
If omitted from other palette types, uses all colours.}

\item{code}{Type of colour code to return. Choices are:
\code{hex}, \code{rgb}, \code{cymk}
If omitted, returns hex code}
}
\value{
A vector of colours.
}
\description{
Returns selected colour palette as a vector of the selected code type (hex/rgb/cymk).
}
\keyword{colours}
