% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_palette.R
\name{focus_colours}
\alias{focus_colours}
\title{List of colour codes in the 'focus' colour palette}
\usage{
focus_colours(code = c("hex", "rgb", "cymk", "names"))
}
\arguments{
\item{code}{type of colour code required. Choices are:
\code{hex}, \code{rgb}, \code{cymk}, \names{names}}
}
\value{
vector of the required type of colour code  from the focus palette
}
\description{
List of colour codes in the 'focus' colour palette
}
