% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_picker.R
\name{palette_picker}
\alias{palette_picker}
\title{Palette selection and return}
\usage{
palette_picker(type, colour_format = "hex")
}
\arguments{
\item{type}{type of palette required (categorical, duo, sequential, focus)}

\item{colour_format}{type of colour format required (hex, rgb), default is hex}
}
\value{
vector of the required type of colour code  from the required palette
}
\description{
Used in the af_colours function. Returns a vector of the required colour codes from the palette selected in the input "type"
}
\examples{
afcolours:::palette_picker(type = "categorical", colour_format = "rgb")
}
